Finish generalizing

+++ downloader/odc.py
@@ -28,7 +28,7 @@ import os
 from pathlib import Path
 import logging
 
-from config import *
+from config import DATACUBE_CONF, GLOBAL_DATA_FOLDER, DATASETS
 from utils import verify_database_connection, ensure_odc_connection_and_database_initialization, check_global_data_folder
 
 logging_config_file = Path(Path(__file__).parent, 'logging.yaml')
@@ -64,6 +64,9 @@ def parse_parameter() -> argparse.Namespace:
     parser.add_argument('-d', '--db',
                         help="Database name. Defaults to 'opendatacube'.",
                         required=False, type=str, default='opendatacube')
+    parser.add_argument('-n', '--no-ping',
+                        help="Do not try to ping the database host before trying to connect to the correct port.",
+                        required=False, type=bool, default=False)
     parser.add_argument('-r', '--max-retries',
                         help="Maximum number of retries while waiting for the database to become available. Defaults to 15.",
                         required=False, type=int, default=15)
@@ -87,14 +90,13 @@ def parse_parameter() -> argparse.Namespace:
     ==============================
     - empty values are allowed
 
-    Download filter
+    Datasets
     ---------------
-    landsat_products_ids    : {}
-    bands                   : {}
-
-    General
-    -------
-
+    {}
+    
+    Dataset loader
+    ---------------
+    {}
 
     OpenDataCube Connection Configuration
     -------------------------------------
@@ -103,38 +105,38 @@ def parse_parameter() -> argparse.Namespace:
     pass        : '{}'
     max retries : '{}'
     sleep       : '{}'
-    """.format(args.landsat_product_ids, args.bands, args.host, args.user,
-               len(args.password) * '*', args.max_retries, args.sleep)
+    """.format([dataset[0] for dataset in DATASETS], [dataset[1] for dataset in DATASETS],
+               args.host, args.user, len(args.password) * '*', args.max_retries, args.sleep))
 
     return args
 
 
-def create_product_metadata_eo3(odc_product_name, metadata_dict):
+def create_product_yaml_eo3(odc_product_name, metadata_dict):
     """
     Create product metadata in eo3 metadata format
 
     :param odc_product_name: product name
     :param metadata_dict:
-    :returns: `dict` of product metadata
+    :return: `dict` of product metadata
     """
 
-    # assert odc_product_name
+    assert odc_product_name == metadata_dict['name'] == metadata_dict['metadata']['product']['name']
 
     measurements = []
-    for band in bands:
+    for band in metadata_dict['bands']:
         measurements.append(
             {
                 'name': band,
-                'dtype': dtype,
-                'units': bands[band]['unit'],
-                'nodata': nodata,
-                'aliases': [bands[band]['aliases']],
+                'dtype': metadata_dict['bands'][band]['dtype'],
+                'units': metadata_dict['bands'][band]['unit'],
+                'nodata': metadata_dict['bands'][band]['nodata'],
+                'aliases': metadata_dict['bands'][band]['aliases'],
             }
         )
 
     product_yaml = {
         'metadata_type': 'eo3',
-        'name': odc_product_name,
+        'name': metadata_dict['name'],
         'description': metadata_dict['description'],
         'metadata': metadata_dict['metadata'],
         'measurements': measurements,
@@ -142,14 +144,16 @@ def create_product_metadata_eo3(odc_product_name, metadata_dict):
     return product_yaml
 
 
-def create_dataset_metadata_eo3(odc_product_name, metadata_dict):
+def create_dataset_yaml_eo3(odc_product_name, metadata_dict):
     """
     Create dataset metadata in eo3 metadata format
 
     :param metadata_dict:
     :param odc_product_name: odc product name
-    :returns: `dict` of dataset metadata
+    :return: `dict` of dataset metadata
     """
+
+    assert odc_product_name == metadata_dict['name']
 
     # Band information
     measurements = {}
@@ -161,14 +165,13 @@ def create_dataset_metadata_eo3(odc_product_name, metadata_dict):
         }
         band_idx = band_idx + 1
 
+    # Define yaml file
     dataset_yaml = {
+        '$schema': metadata_dict.get('schema', 'https://schemas.opendatacube.org/dataset'),
         'id': metadata_dict['id'],
-        '$schema': 'https://schemas.opendatacube.org/dataset',
         'product': {
-            'name': odc_product_name,
+            'name': metadata_dict['product_name'],
         },
-        'keywords': metadata_dict.get('keywords'),
-        'links': metadata_dict.get('links'),
         'crs': metadata_dict['crs'],
         'geometry': {
             'type': 'Polygon',
@@ -190,6 +193,11 @@ def create_dataset_metadata_eo3(odc_product_name, metadata_dict):
         },
         'lineage': metadata_dict.get('lineage')
     }
+
+    # Add additional key-value pairs which are not part of the default metadata
+    if metadata_dict.get('additions'):
+        dataset_yaml.update(metadata_dict['additions'])
+
     return dataset_yaml
 
 
@@ -197,9 +205,9 @@ def add_datasets_to_odc(global_data_folder, dc):
     """
     Add datasets to Open Data Cube index
 
-    :param global_data_folder: folder where where data and metadata yamls are located
+    :param global_data_folder: folder where data and metadata yamls are located
     :param dc: Open Data Cube
-    :returns:
+    :return: None
     """
 
     for dataset in DATASETS:
@@ -211,26 +219,25 @@ def add_datasets_to_odc(global_data_folder, dc):
         Start indexing dataset '{}'
         """.format(dataset[0]))
 
-        # index products and save product yaml files
+        # ToDo: add option to overwrite yaml files?
+
+        # Index products and save product yaml files
         for odc_product in odc_product_names:
             if odc_product not in dc.list_products()['name'].values:
                 product_metadata = loader.create_product_metadata_eo3(odc_product)
-                product_yaml = create_product_metadata_eo3(odc_product, product_metadata)
-                # ToDo: check if yaml files should be saved here
-                # ToDo: where does product_filename come from?
-                product_filename = Path(global_data_folder, '{}.odc-product.yaml'.format(odc_product))
-
-                # ToDo: overwrite option?
+                product_yaml = create_product_yaml_eo3(odc_product, product_metadata)
+                product_filename = os.path.join(global_data_folder,
+                                                loader.get_folder(),
+                                                '{}.odc-product.yaml'.format(odc_product))
                 if not os.path.exists(product_filename):
                     with open(product_filename, 'w') as f:
                         yaml.dump(product_yaml, f, default_flow_style=False, sort_keys=False)
-
                 dc.index.products.add_document(product_yaml)
                 logger.info("Added product family '{}' to the index".format(odc_product))
             else:
                 logger.info("Product family '{}' already in index".format(odc_product))
 
-        # index dataset and save dataset yaml files
+        # Index dataset and save dataset yaml files
         idx_product = 1
         for odc_product in odc_product_names:
             logger.info("[Product {}/{}] Start indexing datasets for product '{}'".format(
@@ -238,24 +245,28 @@ def add_datasets_to_odc(global_data_folder, dc):
                 len(odc_product_names),
                 odc_product))
             idx_dataset = 1
+            # 'odc_dataset' can be a file or a folder (e.g. when bands are in separate files like for Landsat 8)
             for odc_dataset in odc_product_dataset_map[odc_product]:
-
-                # ToDo: how to get id?
-                dataset_id = str(uuid.uuid5(uuid.NAMESPACE_URL, file_name))
-
+                if not os.path.exists(odc_dataset):
+                    logger.info("Dataset '{}' does not exist! Continue with next.".format(odc_dataset))
+                    continue
+                dataset_id = str(uuid.uuid5(uuid.NAMESPACE_URL, odc_dataset))
                 if not dc.index.datasets.get(id_=dataset_id) is None:
                     logger.info("[Dataset {}/{}] Dataset with id '{}' already in the index".format(
                         idx_dataset,
                         len(odc_product_dataset_map[odc_product]),
                         dataset_id))
                 else:
-                    dataset_metadata = loader.create_dataset_metadata_eo3(odc_dataset, odc_product)
-                    dataset_yaml = create_dataset_metadata_eo3(dataset_metadata)
-                    # ToDo: where does dataset_filename come from?
-                    dataset_filename = Path(global_data_folder, '{}.odc-metadata.yaml'.format(file_name)).resolve()
+                    dataset_metadata = loader.create_dataset_metadata_eo3(odc_product, odc_dataset)
+                    dataset_yaml = create_dataset_yaml_eo3(odc_product, dataset_metadata)
+                    # Note: be careful with choosing path and name of dataset yaml file because file or folder
+                    # names might exist more than once (this is the case, e.g., for the AnthroProtect dataset)
+                    # Put dataset yaml file next to the file/folder
+                    file_or_folder_path, file_or_folder_name = os.path.split(os.path.normpath(odc_dataset))
+                    dataset_filename = os.path.join(file_or_folder_path,
+                                                    '{}.odc-metadata.yaml'.format(file_or_folder_name))
                     with open(dataset_filename, 'w') as f:
                         yaml.dump(dataset_yaml, f, default_flow_style=False, sort_keys=False)
-
                     resolver = datacube.index.hl.Doc2Dataset(index=dc.index, eo3=True)
                     dataset = resolver(dataset_yaml, dataset_filename.as_uri())
                     dc.index.datasets.add(dataset[0])
@@ -263,9 +274,9 @@ def add_datasets_to_odc(global_data_folder, dc):
                         idx_dataset,
                         len(odc_product_dataset_map[odc_product]),
                         dataset_id))
-
                 idx_dataset = idx_dataset + 1
             idx_product = idx_product + 1
+    return None
 
 
 def main():
@@ -276,7 +287,7 @@ def main():
     while times_slept < args.max_retries:
         # check "connection" to odc (?) or later
         logger.info("[{}/{}] Check database connection".format(str(times_slept + 1), str(args.max_retries)))
-        if verify_database_connection(args.db, args.host, args.user, args.password, True):
+        if verify_database_connection(args.db, args.host, args.user, args.password, args.no_ping):
             db_conn_ok = True
             break
         time.sleep(args.sleep)
@@ -291,22 +302,19 @@ def main():
                                                       args.password,
                                                       '')
 
-
-    zip_file = DATA_FOLDER + 'anthroprotect.zip'
-    out_folder = DATA_FOLDER + 'anthroprotect'
+    global_data_folder = GLOBAL_DATA_FOLDER
 
     # Check data folder
-    check_global_data_folder(DATA_FOLDER)
+    check_global_data_folder(global_data_folder)
 
     # Download datasets
     for dataset in DATASETS:
-        loader = dataset[0](DATA_FOLDER)
-        loader.download(DATA_FOLDER)
-        unzip_anthroprotect(zip_file, out_folder)
+        loader = dataset[0](global_data_folder)
+        loader.download(global_data_folder, loader.force_download)
 
     # Add products and datasets to Open Data Cube index
     dc = datacube.Datacube(config=DATACUBE_CONF)
-    add_datasets_to_odc(DATA_FOLDER, dc)
+    add_datasets_to_odc(global_data_folder, dc)
 
 
 if __name__ == '__main__':
